+++ tools/apps/dlc-dev-api/src/modules/health/health.module.ts
@@ -1,9 +1,11 @@
 import { Module } from '@nestjs/common';
 import { HealthController } from './health.controller';
+import { MetricsService } from './metrics.service';
 import { AuthModule } from '../auth/auth.module';
 
 @Module({
   imports: [AuthModule],
   controllers: [HealthController],
+  providers: [MetricsService],
 })
 export class HealthModule {}